% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gfe.R
\name{print.summary.gfe}
\alias{print.summary.gfe}
\title{Print a GFE/WGFE Summary}
\usage{
\method{print}{summary.gfe}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.gfe}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Print method for \code{summary.gfe} objects: displays the call, panel dimensions,
coefficient table(s), group proportions and standard deviations, objective value,
and, if \code{ggplot2} is installed, a plot of the \eqn{\alpha}-time series.
}
