% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxFunctions.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round preserve sum.}
\source{
\url{https://www.r-bloggers.com/round-values-while-preserve-their-rounded-sum-in-r/} and \url{http://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}
}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of digits to take in account in the rounding process.}
}
\value{
\code{round_preserve_sum} returns \code{y} with round vector.
}
\description{
Rounds a vector of numbers while preserving the sum of them.
}
\examples{
sum(c(0.333, 0.333, 0.334))
round(c(0.333, 0.333, 0.334), 2)
sum(round(c(0.333, 0.333, 0.334), 2))
round_preserve_sum(c(0.333, 0.333, 0.334), 2)
sum(round_preserve_sum(c(0.333, 0.333, 0.334), 2))
}
